Rewrite to use only one state machine for all ticket control forms
