+++ src/page-modules/contact/index.ts
@@ -1,14 +1,7 @@
-export {
-  ContactPageLayout,
-  type ContactPageLayoutProps,
-  TicketControlPageLayout,
-  type TicketControlPageLayoutProps,
-} from './layouts';
+export { ContactPageLayout, type ContactPageLayoutProps } from './layouts';
 
-export { PostponePaymentForm } from './ticket-control';
 export { RefundForm } from './travel-guarantee';
-export { FeeComplaintForm } from './ticket-control/complaint';
 export { type Line } from './server/journey-planner/validators';
 export { shouldShowContactPage } from './utils';
-export { FeedbackForm } from './ticket-control/feedback';
 export { MeansOfTransportContent } from './means-of-transport';
+export { default as TicketControlPageContent } from './ticket-control';
