+++ src/page-modules/contact/layouts/index.ts
@@ -2,8 +2,3 @@ export {
   default as ContactPageLayout,
   type ContactPageLayoutProps,
 } from './contact-page-layout';
-
-export {
-  default as TicketControlPageLayout,
-  type TicketControlPageLayoutProps,
-} from './ticket-control-page-layout';
