+++ src/page-modules/contact/ticket-control/forms/postponePayment.tsx
@@ -1,32 +1,27 @@
-import { FormEventHandler, useState } from 'react';
-import { Button } from '@atb/components/button';
 import { Input } from '../../components/input';
 import { SectionCard } from '../../components/section-card';
 import { PageText, useTranslation } from '@atb/translations';
-import { useMachine } from '@xstate/react';
-import { postponePaymentForm } from './postponePaymentFormMachine';
+import { ticketControlFormEvents } from '../events';
+import { ContextProps } from '../ticket-controll-form-machine';
+import { Typo } from '@atb/components/typography';
 
-export const PostponePaymentForm = () => {
+type PostponePaymentFormProps = {
+  state: { context: ContextProps };
+  send: (event: typeof ticketControlFormEvents) => void;
+};
+
+export const PostponePaymentForm = ({
+  state,
+  send,
+}: PostponePaymentFormProps) => {
   const { t } = useTranslation();
-  const [state, send] = useMachine(postponePaymentForm);
-
-  // Local state to force re-render to display errors.
-  const [forceRerender, setForceRerender] = useState(false);
-
-  const onSubmit: FormEventHandler<HTMLFormElement> = async (e) => {
-    e.preventDefault();
-    send({ type: 'VALIDATE' });
-
-    // Force a re-render with dummy state.
-    if (Object.keys(state.context.errorMessages).length > 0) {
-      setForceRerender(!forceRerender);
-    }
-  };
 
   return (
-    <form onSubmit={onSubmit}>
+    <div>
       <SectionCard title={PageText.Contact.ticketControl.postponePayment.title}>
-        <p>{t(PageText.Contact.ticketControl.postponePayment.info)}</p>
+        <Typo.p textType="body__primary">
+          {t(PageText.Contact.ticketControl.postponePayment.info)}
+        </Typo.p>
 
         <Input
           label={PageText.Contact.inputFields.feeNumber.label}
@@ -110,12 +105,7 @@ export const PostponePaymentForm = () => {
           }
         />
       </SectionCard>
-      <Button
-        title={t(PageText.Contact.submit)}
-        mode={'interactive_0--bordered'}
-        buttonProps={{ type: 'submit' }}
-      />
-    </form>
+    </div>
   );
 };
