--- src/page-modules/contact/ticket-control/index.ts
@@ -1,1 +1,0 @@
-export { PostponePaymentForm } from './postpone-paymnet';
